Removed all reagent/re-frame.
This is supposed to be a demo. Re-frame is really only cluttering things
up. It definitely helps keep a large app in check, but for small demo
purposes it's really getting in the way. Look at this, 400 lines of
framework code vs less than 300 lines of demo code.

I knew this already I suppose, but it really has to be driven home sometimes.